chore: refactor code fo the dropdown button
