+++ services/frontend/web/src/assets/locales/en.json
@@ -5,5 +5,7 @@
 	"Profile": "Profile",
 	"Quick Match": "Quick Match",
 	"Game": "Game",
-	"Leaderboard": "Leaderboard"
+	"Leaderboard": "Leaderboard",
+	"Account Settings": "Account Settings",
+	"Edit Profile": "Edit Profile"
 }
