+++ services/frontend/web/src/assets/locales/es.json
@@ -5,5 +5,7 @@
 	"Profile": "Perfil",
 	"Quick Match": "Parte Rápida",
 	"Game": "Juego",
-	"Leaderboard": "Junta de líderes"
+	"Leaderboard": "Junta de líderes",
+	"Account Settings": "Configuraciones de la cuenta",
+	"Edit Profile": "Editar perfil"
 }
