+++ services/frontend/web/src/assets/locales/fr.json
@@ -5,6 +5,7 @@
 	"Profile": "Profil",
 	"Quick Match": "Partie Rapide",
 	"Game": "Jeu",
-	"Leaderboard": "Board des leaders"
-
+	"Leaderboard": "Board des leaders",
+	"Account Settings": "Parametres du compte",
+	"Edit Profile": "Modifier le profil"
 }
