+++ services/frontend/web/src/home.ts
@@ -1,19 +1,10 @@
 import './style.css';
-import { t, changeLanguage } from './translation';
-import { currentLang } from './translation';
+import { t } from './translation';
+import { renderLanguageDropdownButton } from './languageDropdownButton'
 
 export function renderHome(main: HTMLElement) {
   main.innerHTML = `
-      <div class="dropdown">
-        <button id="language-switch" class="px-4 py-2 rounded bg-gray-200 hover:bg-gray-300 text-black font-semibold">
-          ${t(currentLang)}  <!-- Show current language name -->
-        </button>
-        <div id="language-options" class="dropdown-content hidden absolute bg-white border rounded shadow-md mt-1">
-          <div class="dropdown-item px-4 py-2 hover:bg-gray-200 cursor-pointer" data-lang="en">English</div>
-          <div class="dropdown-item px-4 py-2 hover:bg-gray-200 cursor-pointer" data-lang="fr">Français</div>
-          <div class="dropdown-item px-4 py-2 hover:bg-gray-200 cursor-pointer" data-lang="es">Español</div>
-        </div>
-      </div>
+	<div class="top-center-button" id="lang-dropdown-container"></div>
       <div class="min-h-screen flex items-center justify-center bg-white">
         <a
           href="/central"
@@ -29,34 +20,8 @@ export function renderHome(main: HTMLElement) {
     </div>
   `;
 
-function setupLanguageDropdown(main: HTMLElement) {
-  const switchBtn = document.getElementById('language-switch');
-  const options = document.getElementById('language-options');
-
-  if (!switchBtn || !options) return;
-
-  switchBtn.addEventListener('click', () => {
-    options.classList.toggle('hidden');
-  });
-
-  options.querySelectorAll('.dropdown-item').forEach(item => {
-    item.addEventListener('click', () => {
-      const lang = (item as HTMLElement).dataset.lang;
-      if (lang && lang !== currentLang) {
-        changeLanguage(lang);
-        options.classList.add('hidden');
-        renderHome(main);
-      }
-    });
-  });
-
-  document.addEventListener('click', (e) => {
-    if (!main.contains(e.target as Node)) {
-      options.classList.add('hidden');
-    }
-  });
-}
-
-setupLanguageDropdown(main);
-
+    const langDropdown = document.getElementById('lang-dropdown-container');
+  if (langDropdown) {
+    renderLanguageDropdownButton(langDropdown, () => renderHome(main));
+  }
 }
